Break up TaskList into separate classes of Commands and  AddTask
