+++ src/main/java/ListCommands/SamException.java
@@ -1,6 +1,8 @@
+package ListCommands;
+
 // Custom exception class for Sam application
 public class SamException extends Exception {
-    // Constructor to create a SamException with an error message
+    // Constructor to create a Commands.SamException with an error message
     public SamException(String errorMessage) {
         if (errorMessage != null) {
             // Print the error message if not null
